+++ misc/update_DESCRIPTION_NEWS.R
@@ -370,6 +370,9 @@ my_news$add_bullet(c("utils", "grid"))
 my_desc$set_dep("utils", type = desc::dep_types[1])
 my_desc$set_dep("grid", type = desc::dep_types[1])
 
+my_news$add_subtitle("Bugfixes")
+my_news$add_bullet(c("changing error messages"))
+
 
 # save everything ---------------------------------------------------------