+++ tests/testthat/test-checkdir.R
@@ -2,16 +2,10 @@ context("check checkdir")
 
 test_that("NA, NULL is handeld right", {
 
-  old_lang <- Sys.getenv("LANG")
-
-  Sys.setenv(LANG = "en")
-
-  on.exit(Sys.setenv(LANG = old_lang))
-
   expect_error(checkdir(path = NA, recursive = FALSE),
-               "invalid 'path' argument")
+               "path is NA or NULL")
   expect_error(checkdir(path = NULL, recursive = FALSE),
-               "argument is of length zero")
+               "path is NA or NULL")
 
 })